feat(web): replace sign-in and sign-up pages with modals and redirect on sign-in
