+++ apps/web/app/auth/sign-up/[[...sign-up]]/page.tsx
@@ -1,25 +1,22 @@
 'use client';
-import Logo from '@/components/shared/branding/logo';
-import SignUpForm from '@/components/specific/auth/clerk/form/signup-form';
-import VerificationForm from '@/components/specific/auth/clerk/form/verification-form';
-import { useAuthFlowStore } from '@/components/specific/auth/clerk/stores/auth-flow-store';
-import { Card, CardHeader, CardContent } from '@hwei/ui/shadcn/card';
-import React from 'react';
+
+import SignUpModal from '@/components/specific/auth/clerk/modals/sign-up-modal';
+import { useAuth } from '@clerk/nextjs';
+import { useRouter } from 'next/navigation';
+import { useEffect } from 'react';
 
 const SignUpPage = () => {
-	const { step } = useAuthFlowStore();
+	const { isSignedIn } = useAuth();
+	const router = useRouter();
 
+	useEffect(() => {
+		if (isSignedIn) {
+			router.push('/');
+		}
+	}, [isSignedIn]);
 	return (
 		<div className="full relative">
-			<Card className="h-full w-full px-12 py-10 shrink-0 lg:min-w-[36rem] overflow-hidden">
-				<CardHeader className="col-center">
-					<Logo />
-				</CardHeader>
-
-				<CardContent className="mt-4 mb-6 col-center">
-					{step === 'verification' ? <VerificationForm /> : <SignUpForm />}
-				</CardContent>
-			</Card>
+			<SignUpModal />
 		</div>
 	);
 };
